requirements.txt
pandas
gradio
flaml
matplotlib
seaborn
Pillow
joblib
scikit-learn